Refactor menu system: replace Menu class with MainMenu class for improved structure and functionality
