+++ StorageOffice/classes/Logic/MenuHandler.cs
@@ -8,7 +8,7 @@ public class MenuHandler
     public static void MainMenu()
     {
         
-        var menu = new Menu("Storage System manager", "Welcome", new RadioSelect(new List<RadioOption>
+        var menu = new MainMenu("Storage System manager", "Welcome", new RadioSelect(new List<RadioOption>
         {
             new RadioOption("Option 1"),
             new RadioOption("Option 2"),
